remove old qnns
