+++ test/algorithms/classifiers/test_vqc.py
@@ -171,7 +171,7 @@ class TestVQC(QiskitMachineLearningTestCase):
         self.assertGreater(score, 0.5)
 
     def _get_num_classes(self, func):
-        """Wrapper to record the number of classes assumed when building CircuitQNN."""
+        """Wrapper to record the number of classes assumed when building SamplerQNN."""
 
         @functools.wraps(func)
         def wrapper(num_classes):
@@ -207,7 +207,7 @@ class TestVQC(QiskitMachineLearningTestCase):
         with self.subTest("Test all batches assume the correct number of classes."):
             self.assertTrue((np.asarray(num_classes_list) == 3).all())
 
-        with self.subTest("Check correct number of classes is used to build CircuitQNN."):
+        with self.subTest("Check correct number of classes is used to build SamplerQNN."):
             self.assertTrue((np.asarray(self.num_classes_by_batch) == 3).all())
 
     def test_multilabel_targets_raise_an_error(self):
